Moves/Renames akka-http related files to be on par with netty and copies the gzip decoder actually this marks akka-http backend as non experimental only depends on the akka-http-core model by using the compressor of akka-stream
